# Saas-Project/backend/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS UMKM Backend API
  description: |
    API documentation for Indonesian SME licensing and business management platform.
    This API provides endpoints for user management, company registration, license 
    management, and business operations.
  version: 0.1.0
  contact:
    name: SaaS UMKM Team
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api-staging.saas-umkm.com/api
    description: Staging server
  - url: https://api.saas-umkm.com/api
    description: Production server
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account
      operationId: registerUser
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: User login
      description: Authenticates a user and returns a JWT token
      operationId: loginUser
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    UserRegistrationRequest:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecureP@ssw0rd
        full_name:
          type: string
          example: John Doe
        phone:
          type: string
          example: '+62812345678'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: SecureP@ssw0rd
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        full_name:
          type: string
          example: John Doe
        role:
          type: string
          enum: [umkm_owner, admin_staff, super_admin]
          example: umkm_owner
        status:
          type: string
          enum: [active, pending_verification, suspended]
          example: active
        phone:
          type: string
          nullable: true
          example: '+62812345678'
        email_verified:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: '2025-07-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-07-01T12:00:00Z'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid input
        message:
          type: string
          example: Email is already in use
        status:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: '2025-07-01T12:00:00Z'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT